prand decomposed from pcompat
